requirements.txt
requests
pywebview
google-genai
psutil
Pillow
zstandard
